chore: refactor "session" naming to "token"
