+++ packages/core/src/_exports/index.ts
@@ -38,5 +38,5 @@ export {getSessionStore} from '../auth/getSessionStore'
 export {getSidUrlSearch} from '../auth/sessionId'
 export type {LoggedInState, SessionState, SessionStore} from '../auth/sessionStore'
 export {LOGGED_IN_STATES} from '../auth/sessionStore'
-export {tradeTokenForSession} from '../auth/tradeTokenForSession'
+export {tradeSessionIdForToken} from '../auth/tradeSessionIdForToken'
 export type {CurrentUser, Role} from '@sanity/types'
