+++ packages/core/src/auth/fetchCurrentUser.ts
@@ -8,17 +8,17 @@ import {getSessionStore} from './getSessionStore.ts'
  * Fetches the current user from the Sanity API.
  * @internal
  * @param sanityInstance - The Sanity instance.
- * @param sessionId - The session ID.
  * @returns The current user or null if the user is not logged in.
  */
-export const fetchSessionUser = async (
+export const fetchCurrentUser = async (
   sanityInstance: SanityInstance,
-  sessionId?: string | null,
 ): Promise<CurrentUser | null> => {
-  sessionId = sessionId ?? getSessionStore(sanityInstance).getState().sessionId
-  // TODO: handle case where sessionId is null
+  const token = sanityInstance.config.token
+  if (!token) {
+    throw new Error('No token is set for this instance')
+  }
   const client = getClient({apiVersion: 'v2024-11-22'}, sanityInstance)
-  const user = await client.withConfig({token: sessionId!}).request<CurrentUser>({
+  const user = await client.withConfig({token}).request<CurrentUser>({
     method: 'GET',
     uri: `/users/me`,
   })
